Death screen loads
When you die the death screen now loads, and the lightning gun as a hit animation and sound

+++ Characterprototype/Assets/LoadScore.cs
@@ -8,7 +8,7 @@ public class LoadScore : MonoBehaviour
     public Text text;
     void Start()
     {
-        text.text = PlayerPrefs.GetInt("Score").ToString();
+        text.text = PlayerPrefs.GetFloat("Score").ToString();
     }
